refactor: move environment and API key config to environment variables only
