+++ example-dimo-auth/src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './Input';
-export * from './Select';
 export * from './ConfigForm';
