+++ example-dimo-auth/src/constants/storage.ts
@@ -1,0 +1,1 @@
+export const STORAGE_KEY = 'dimoConfig';
