+++ example-dimo-auth/src/types/dimo.ts
@@ -1,0 +1,6 @@
+export interface DimoConfig {
+  clientId: string;
+  redirectUri: string;
+  environment: 'production' | 'development';
+  apiKey: string;
+}
